Changes around error and warning behaviour `optional`/`optionalAt`
Previously; using `optional` or `optionalAt` would succeed quite happily with
the fallback value, _even_ if one of the structures on the path wasn't even an
object.

Now, such a thing _does_ result in an error (Expected an object). As a
side-effect, this allows marking structures that were traversed as having been
used, which is important when using `optional` on an empty object: decoding
succeeded and we know it's an object, so we can *and should* mark it as used.

Thanks to @michaeljones for surfacing this issue.

Resolves #3